% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{bindis}
\alias{bindis}
\title{Binary to discrete response matrix}
\usage{
bindis(y, n_alts)
}
\arguments{
\item{y}{Matrix containing the binary choice data.}

\item{n_alts}{The number of alternatives per choice set.}
}
\value{
A matrix with discrete values, indicating the choices.
}
\description{
Transforms a matrix with binary choice data for each respondent (columns),
to a matrix with discrete values representing the choices.
}
