% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{CharBin}
\alias{CharBin}
\title{Character to binary responses.}
\usage{
CharBin(resp, resp.opt, n.alts, no.choice = FALSE)
}
\arguments{
\item{resp}{String vector containing input responses}

\item{n.alts}{The number of alternatives per choice set.}

\item{no.choice}{Logical value indicating whether a no.choice option is
provided or not. Default = FALSE.}

\item{resp.options}{String vector containing all possible responses. The
response options should be specified in increasing order, starting with the
no choice option (if included).}
}
\value{
A binary response vector.
}
\description{
Transforms character input responses to binary response vector.
}
\examples{
rm(list=ls())
resp.opt <- c("no.choice", "alt1", "alt2", "alt3")
resp <- c("alt1", "alt3", "alt2", "no.choice", "alt1") 
CharBin(resp = resp, resp.opt = resp.opt, n.alts = 3, no.choice = TRUE)
}
