% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{d_err}
\alias{d_err}
\title{D-error}
\usage{
d_err(par, des, n_alts)
}
\arguments{
\item{par}{Vector containing parameter values.}

\item{des}{A design matrix in which each row is a profile.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}
}
\value{
D-error.
}
\description{
Function to calculate d error given a design, and parameter values.
}
