% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{DBerrS}
\alias{DBerrS}
\title{Sequential DB-error}
\usage{
DBerrS(full.comb, cand.set, par.samples, des, n.alts, cte.des, i.cov, n.par,
  weights)
}
\arguments{
\item{full.comb}{A matrix with on each row a possible combination of
profiles.}

\item{cand.set}{A numeric matrix in which each row is a possible profile. The
\code{\link{Profiles}} function can be used to generate this.}

\item{par.samples}{A matrix in which each row is a sample from the
multivariate prior parameter distribution.}

\item{des}{A design matrix in which each row is a profile.}

\item{n.alts}{Numeric value indicating the number of alternatives per choice
set.}

\item{cte.des}{A matrix which represent the alternative specific constants.
If there are none it value is \code{NULL}.}

\item{i.cov}{Inverse of covariance matrix.}

\item{n.par}{Number of parameters.}
}
\value{
The DB errors of the designs in which each design is a combination
with of the initial design with a potential choice set.
}
\description{
Function to calculate DB-errors for potential choice sets in combination with
an initial design.
}
