% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{DerrS}
\alias{DerrS}
\title{Sequential D-error}
\usage{
DerrS(par.samples, set, des, n.alts, i.cov, n.par)
}
\arguments{
\item{par.samples}{A matrix in which each row is a sample from the
multivariate prior parameter distribution.}

\item{set}{A choice set in which each row is a profile.}

\item{des}{A design matrix in which each row is a profile.}

\item{n.alts}{Numeric value indicating the number of alternatives per choice
set.}

\item{i.cov}{Inverse of covariance matrix.}

\item{n.par}{Number of parameters.}
}
\description{
Function to calculate D-errors if set would be part of design.
}
