% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{design}
\alias{design}
\title{Discrete choice design.}
\format{An object of class \code{matrix} with 16 rows and 6 columns.}
\usage{
data(design)
}
\description{
This discrete choice design is generated using the \code{\link{Modfed}}
function. There are 8 choice sets, each containig 2 alternatives (rows). The
alternatives consist of 3 attributes (time, price, comfort) with each 3
levels, all of which are dummy coded (columns).
}
\keyword{datasets}
