% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_design}
\alias{example_design}
\title{Discrete choice design.}
\format{A matrix with 16 rows and 6 columns.}
\usage{
data(example_design)
}
\description{
This discrete choice design is generated using the \code{\link{Modfed}}
function. There are 8 choice sets, each containig 2 alternatives (rows). The
alternatives consist of 3 attributes (time, price, comfort) with 3
levels each, all of which are dummy coded (columns).
}
\keyword{data}
