% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_design2}
\alias{example_design2}
\title{Discrete choice design.}
\format{A matrix with 24 rows and 8 columns.}
\usage{
data(example_design2)
}
\description{
This discrete choice design is generated using the \code{\link{Modfed}}
function. There are 8 choice sets, each containig 3 alternatives (rows). The
alternatives consist of 3 attributes (time, price, comfort) with each 3
levels, all of which are dummy coded (columns). The first two colums are
alternative specific constants for alternative 1 and 2.
}
\keyword{data}
