% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{full_sets}
\alias{full_sets}
\title{All choice sets}
\usage{
full_sets(candset, n_alts)
}
\arguments{
\item{candset}{A numeric matrix in which each row is a possible profile.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}
}
\value{
Matrix with all possible combinations of profiles (choice sets).
}
\description{
Generates all possible combinations of choice sets.
}
