% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{g_dens}
\alias{g_dens}
\title{Density multivariate t-distribution}
\usage{
g_dens(par, g_mean, g_covar)
}
\arguments{
\item{par}{Numeric vector with parametervalues.}

\item{g_mean}{vector containing the mean of the multivariate t-distribution.}

\item{g_covar}{covariance matrix of the multivariate t-distribution.}
}
\value{
density
}
\description{
Density multivariate t-distribution
}
