% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{Gdens}
\alias{Gdens}
\title{Density multivariate t-distribution}
\usage{
Gdens(par, g.mean, g.covar)
}
\arguments{
\item{par}{Numeric vector with parametervalues.}

\item{g.mean}{vector containing the mean of the multivariate t-distribution.}

\item{g.covar}{covariance matrix of the multivariate t-distribution.}
}
\value{
density
}
\description{
Density multivariate t-distribution
}
