% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{hessian}
\alias{hessian}
\title{Hessian}
\usage{
hessian(par, des, covar, n_alts)
}
\arguments{
\item{par}{Numeric vector with parametervalues.}

\item{des}{A design matrix in which each row is a profile.}

\item{covar}{The covariance matrix.}

\item{n_alts}{The number of alternatives in each choice set.}
}
\value{
the hessian matrix
}
\description{
Hessian
}
