% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{imp_sampling}
\alias{imp_sampling}
\title{Importance sampling}
\usage{
imp_sampling(prior_mean, prior_covar, des, n_alts, Y, m, b = 2, ...)
}
\arguments{
\item{prior_mean}{Numeric vector which is the mean of the multivariate normal distribution (prior).}

\item{prior_covar}{A matrix, the covariance matrix of the multivariate normal distribution (prior).}

\item{des}{A design matrix in which each row is a profile.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{Y}{A binary response vector.}

\item{m}{Numeric value. Number of samples = base^m.}

\item{b}{Numeric value indicating the base (default = 2).}
}
\value{
A list containing samples, their associated weights, the maximum likelihood estimates and the estimated covariance matrix.
}
\description{
This functions samples from an imortance density (multivariate t-distribution),
and gives weightes to the samples according to the posterior distribution. The prior is
a normal distribution.
}
