% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{info_design}
\alias{info_design}
\title{Fisher Information of design}
\usage{
info_design(par, des, n_alts)
}
\arguments{
\item{par}{A vector containing the parameter values}

\item{des}{A design matrix in which each row is a profile.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}
}
\value{
Fisher Information matrix.
}
\description{
Returns the Fisher Information of a design, given parameter values
}
