% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{InfoDes}
\alias{InfoDes}
\title{Fisher Information of design}
\usage{
InfoDes(par, des, n.alts)
}
\arguments{
\item{par}{A vector containing the parameter values}

\item{n.alts}{Numeric value indicating the number of alternatives per choice
set.}
}
\value{
Fisher Information matrix.
}
\description{
Returns the Fisher Information of a design, given parameter values.
}
