% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{KL_info}
\alias{KL_info}
\title{Kullback-Leibler information criterion}
\usage{
KL_info(fp, fcomb, par.samples, weights, n.alts, print = FALSE)
}
\arguments{
\item{fp}{A Matrix containing every possible profile (combination of attributelevels). Can be generated by using the "profiles" function.}

\item{fcomb}{A matrix containing in each row the to be considered profile combinations (choice sets). Can be generated by using the "full_sets" function.}

\item{par.samples}{A matrix in which each row is a sample from a multivariate parameter distribution.}

\item{weights}{A vector containing the weights of all the samples.}

\item{n.alts}{Numeric value indicating the number of alternatives per choice set.}

\item{print}{Logical statement indicating whether to print the KL information or not.}
}
\value{
The most efficient choice set, based on the KL information criterion.
}
\description{
Calcultates the most effecient choice set by maximizing the Kullback-Leibler distance.
}
