% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{KLs}
\alias{KLs}
\title{KL information}
\usage{
KLs(full.comb, par.samples, cte.des, cand.set, weights)
}
\arguments{
\item{full.comb}{A matrix in which each row is a possible combination of profiles.}

\item{par.samples}{A matrix in which each row is a sample from the
multivariate parameter distribution. See also \code{\link{ImpsamplingMNL}}.}

\item{cand.set}{A numeric matrix in which each row is a possible profile. The
\code{\link{Profiles}} function can be used to generate this.}

\item{weights}{A vector containing the weights of the samples. Default is \code{NULL}, See also \code{\link{ImpsamplingMNL}}.}
}
\value{
Numeric value indicating the Kullback-Leibler divergence.
}
\description{
Calculates the Kullback-Leibler divergence for all posible choice sets, given
parameter values.
}
