% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{lattice}
\alias{lattice}
\title{Lattice multivariate standard normal distribution.}
\usage{
lattice(K, b, m)
}
\arguments{
\item{K}{Numeric value indicating the dimensionality of the grid.}

\item{b}{Numeric value indicating the base.}

\item{m}{Numeric value. Number of samples=b^m.}
}
\value{
Matrix of lattice points drawn from a multivariate standard normal distribution. Each row is a sample.
}
\description{
Generates a grid of points coming from a multivariate standard normal distribution.
}
