% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{Lattice_mvn}
\alias{Lattice_mvn}
\title{Lattice multivariate normal distribution.}
\usage{
Lattice_mvn(mean, cvar, m, b = 2)
}
\arguments{
\item{mean}{Numeric vector indicating the multivariate mean.}

\item{cvar}{A matrix which specifies the covariance matrix.}

\item{m}{Numeric value. Number of samples = b^m.}

\item{b}{Numeric value indicating the base (default = 2).}
}
\value{
Matrix of lattice points drawn from a multivariate normal distribution. Each row is a sample.
}
\description{
Generates a grid of points coming from a multivariate normal distribution.
}
