% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{Lik}
\alias{Lik}
\title{Likelihood function}
\usage{
Lik(par, des, n.alts, y)
}
\arguments{
\item{par}{Numeric vector with parametervalues.}

\item{des}{A design matrix in which each row is a profile.}

\item{n.alts}{The number of alternatives in each choice set}

\item{y}{A binary response vector.}
}
\value{
the likelihood
}
\description{
Likelihood function
}
