% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{LoadData}
\alias{LoadData}
\title{Load numeric choice data from directory}
\usage{
LoadData(data.dir, type)
}
\arguments{
\item{data.dir}{A character string containing the directory to read from.}

\item{type}{Character vector containing either num or char.}
}
\value{
A data frame containg the full design and all the responses of the
combined data files that were found. Different files are indicated by an
ID variable.
}
\description{
Reads all individual choice data files, created by \code{\link{SurveyApp}}
function, from a directory and concatenates
those files into a single data file. Files containing either "num" or "char"
will be read, with num indicating numeric data and char indicating character
data. For more information, see output of \code{\link{SurveyApp}}.
}
