% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{loaddrop}
\alias{loaddrop}
\title{Load from dropbox}
\usage{
loaddrop(dir)
}
\arguments{
\item{inputDir}{A dropbox directory that contains the design.}
}
\value{
the file in that directory (or concatenated files).
}
\description{
Load design from dropbox map
}
