% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{LogPost}
\alias{LogPost}
\title{log Posterior}
\usage{
LogPost(par, prior.mean, prior.covar, des, n.alts, y)
}
\arguments{
\item{par}{Numeric vector with parametervalues.}

\item{prior.mean}{vector containing the prior mean.}

\item{prior.covar}{matrix containing the prior covariance.}

\item{des}{A design matrix in which each row is a profile.}

\item{n.alts}{The number of alternatives in each choice set.}

\item{y}{A binary response vector.}
}
\value{
the logposterior probability
}
\description{
Calculates the logposterior with a normal prior density
}
