% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{map_resp}
\alias{map_resp}
\title{Transform responses}
\usage{
map_resp(resp, resp_options, n_alts, neutral = T)
}
\arguments{
\item{resp}{String vector containing input responses}

\item{resp_options}{String vector containing all possible responses.
The response options should be specified in increasing order, starting with the neutral response (if included).}

\item{n_alts}{The number of alternatives per choice set.}

\item{neutral}{Logical value indicating whether a neutral option is provided or not. Default = TRUE.}
}
\value{
A binary response vector.
}
\description{
Transforms character input responses to binary response vector.
}
