% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{mindiff}
\alias{mindiff}
\title{Minimum difference choice sets}
\usage{
mindiff(candset, fcomb, lvls, mindiff)
}
\arguments{
\item{candset}{A numeric matrix in which each row is a possible profile.}

\item{fcomb}{A matrix with all possible combinations of profiles (choice sets).}
}
\value{
Matrix with all possible combinations of profiles, taking into account the minimum difference between profiles per choice set.
}
\description{
Filters all possible combinations of choice sets,
taking into account the minimum ammount of attributelevels that need to be different per choice set.
}
