% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_algorithms.R
\name{modfed.db}
\alias{modfed.db}
\title{DB modified federov algorithm}
\usage{
modfed.db(candset, n_sets, n_alts, par_samples, max_iter = Inf)
}
\arguments{
\item{candset}{A numeric matrix in which each row is a possible profile.}

\item{n_sets}{Numeric value indicating the number of choice sets.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{par_samples}{A matrix in which each row is a sample from the multivariate prior parameter distribution.}

\item{max_iter}{A numeric value indicating the maximum number allowed iterations.}
}
\value{
An efficient design, and the associated DB-error.
}
\description{
Modified federov algorithm. Swipes every profile of an initial design with candidate profiles.
By doing this it tries to minimize the DB error. The DB error is te mean D-error over samples
from the prior parameter distribution.
}
