% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nochoice_design}
\alias{nochoice_design}
\title{Discrete choice design with no choice option.}
\format{
A matrix with 24 rows and 7 variables
}
\usage{
data(nochoice_design)
}
\description{
This discrete choice design is generated using the \code{\link{Modfed}}
function. There are 8 choice sets, each containig 3 alternatives (rows), of
which one is a no choice option. The no choice option consist of an
alternative specific constant and zero's for all other attribute levels. There
are three attributes (time, price, comfort) with 3 levels each, all of which
are dummy coded (columns).
}
\keyword{data}
