% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_support.R
\name{respond}
\alias{respond}
\title{Response generation}
\usage{
respond(par, set, n_alts, bin = TRUE)
}
\arguments{
\item{par}{Vector containing parameter values.}

\item{set}{A numeric matrix which represents a choice set. Each row is a profile.}

\item{bin}{Indicates whether the returned value should be a binary vector or a discrete value which denotes the chosen alternative.}
}
\value{
Binary response vector or discrete value indicating the choosing alternative.
}
\description{
Function to generate responses given parameter values and a design matrix according to MNL model.
}
