% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_algorithms.R
\name{seqfed.db}
\alias{seqfed.db}
\title{DB sequential set}
\usage{
seqfed.db(des, candset, n_alts, par_samples, weights, prior_covar)
}
\arguments{
\item{des}{A design matrix in which each row is a profile.}

\item{candset}{A numeric matrix in which each row is a possible profile.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{par_samples}{A matrix in which each row is a sample from the multivariate parameter distribution.}

\item{weights}{A vector containing the weights of the samples.}

\item{prior_covar}{Covariance matrix of the prior distribution.}
}
\value{
The most DB efficient choice set.
}
\description{
Adds the most DB efficient choice set to a design, given (updated) parameter values.
}
