% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregate_design}
\alias{aggregate_design}
\title{Discrete choice aggregate design.}
\format{A matrix with 112 rows and 9 variables}
\usage{
data(aggregate_design)
}
\description{
The dataset contains fictional data for seven participants, each
responding to eight choice sets with two alternatives. Each alternative
consists of three attributes, and each attribute contains three levels, which
are dummy coded.
}
\keyword{data}
