% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{BinDis}
\alias{BinDis}
\title{Binary to discrete choice matrix.}
\usage{
BinDis(y, n.alts, no.choice)
}
\arguments{
\item{y}{NUmeric matrix containing the binary choice data. Each column is a
different ID.}

\item{n.alts}{Numeric value indicating the number of alternatives per choice
set.}

\item{no.choice}{Logical value indicating whether a no choice response could
be observed. This would be a \code{0} for each alternative.}
}
\value{
A matrix with discrete values, indicating the chosen alternatives per
ID.
}
\description{
Transforms a numeric matrix with binary choice data for each respondent
(columns), to a matrix with discrete values representing the chosen
alternatives.
}
\examples{
 
# Binary response data, 2 participants
y <- matrix(data = c(0,1,1,0,0,0,0,1), ncol = 2, byrow = FALSE)
# no choice = TRUE 
BinDis(y = y, n.alts = 2, no.choice = TRUE)
}
