% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_support.R
\name{datatrans}
\alias{datatrans}
\title{data transformation.}
\usage{
datatrans(pkg, des, Y, n_alts, n_sets, n_resp, n_beta, bin)
}
\arguments{
\item{pkg}{Indicates the required package or estimation (1=bayesm, 2=ChoiceModelR, 3=RSGHB, 4=Mixed Probit estimation (Bayesm))}

\item{des}{A design matrix in which each row is a profile.}

\item{Y}{A numeric matrix. Each column represents a respondent, there are n_sets rows (discrete choice data), or n_sets*n_alts rows (binary data).}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{n_sets}{Numeric value indicating the number of choide sets.}

\item{n_resp}{Numeric value indicating the number of respondents.}

\item{n_beta}{Numeric value indicating the number of model parameters that needs to be estimated.}

\item{bin}{Logical value indicating whether the reponse matrix contains binary data (TRUE) or not (FALSE).}
}
\value{
The data ready to be used by the specified package.
}
\description{
Transforms the data into the neccesary format in order to use the bayesm package, ChoiceModelR package or the RSGHB package.
}
