% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{Hessian}
\alias{Hessian}
\title{Hessian}
\usage{
Hessian(par, des, covar, n.alts)
}
\arguments{
\item{par}{Numeric vector with parametervalues.}

\item{des}{A design matrix in which each row is a profile.}

\item{covar}{The covariance matrix.}

\item{n.alts}{The number of alternatives in each choice set.}
}
\value{
the hessian matrix
}
\description{
Hessian
}
