% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idefix.R
\docType{package}
\name{idefix-package}
\alias{idefix}
\alias{idefix-package}
\title{idefix: efficient designs for discrete choice experiments.}
\description{
Generates efficient designs for discrete choice experiments based on the
Multinomial Logit (MNL) model, and individually adapted designs for the Mixed
Multinomial Logit model. The (adaptive) designs can be presented on screen and
choice data can be gathered using a shiny application.
}
\details{
\itemize{
\item To generate efficient designs using the Modified Federov algorithm, please consult the \link[idefix]{Modfed} documentation.
\item To generate efficient designs using the Coordinate Exchange algorithm,
please consult the \link[idefix]{CEA} documentation.
\item To generate adaptive designs using the Modified Fedorov algorithm,
please consult the \link[idefix]{SeqMOD} documentation.
\item To generate adaptive designs using the Coordinate Exchange algorithm,
please consult the \link[idefix]{SeqCEA} documentation.
\item To generate a discrete choice survey on screen, please consult the \link[idefix]{SurveyApp} documentation.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/traets/idefix}
}

}
\author{
\strong{Maintainer}: Frits Traets \email{frits.traets@kuleuven.be}

Other contributors:
\itemize{
  \item Daniel Gil \email{danielgils@gmail.com} [contributor]
}

}
