% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_criterions.R
\name{KL}
\alias{KL}
\title{KL information}
\usage{
KL(set, par_samples, weights)
}
\arguments{
\item{set}{Numeric matrix in which each row is a profile.}

\item{par_samples}{A matrix in which each row is a sample.}

\item{weights}{A vector containing the weights of the samples.}
}
\value{
The Kullback-Leibler divergence.
}
\description{
Calculates the Kullback-Leibler divergence for a choice set, given parameter values.
}
