% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_algorithms.R
\name{KL_select}
\alias{KL_select}
\title{KL set selecting}
\usage{
KL_select(lvls, n.sets, n.alts, par.samples, weights)
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{n.sets}{Numeric value indicating the number of choice sets.}

\item{n.alts}{Numeric value indicating the number of alternatives per choice set.}

\item{par.samples}{A matrix in which each row is a sample.}

\item{weights}{A vector containing the weights of the samples.}
}
\value{
Choice set that maximizes the expected KL divergence.
}
\description{
Provides the set that maximizes the Kullback-Leibler divergence, given parameter values.
}
