% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{logPost}
\alias{logPost}
\title{log Posterior}
\usage{
logPost(par, prior_mean, prior_covar, des, n_alts, Y)
}
\arguments{
\item{par}{Numeric vector with parametervalues.}

\item{prior_mean}{vector containing the prior mean.}

\item{prior_covar}{matrix containing the prior covariance.}

\item{des}{A design matrix in which each row is a profile.}

\item{n_alts}{The number of alternatives in each choice set.}

\item{Y}{A binary response vector.}
}
\value{
the logposterior probability
}
\description{
Calculates the logposterior with a normal prior density
}
