% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{Present}
\alias{Present}
\title{Attribute level choice set}
\usage{
Present(set, lvl.names, coding, intercept = FALSE)
}
\arguments{
\item{set}{A numeric matrix which represents a choice set. Each row is a profile.}

\item{lvl.names}{A list containing the values of each level of each attribute.}

\item{coding}{Type of coding used in the given set. See ?contrasts for more info.}

\item{intercept}{Logical argument indicating whether an intercept is included. The default is False.}
}
\value{
A character matrix which represents the choice set.
}
\description{
Transforms a coded choice set into a choice set containing the attribute levels.
}
