% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{profiles}
\alias{profiles}
\title{Profiles generation}
\usage{
profiles(lvls, coding, intercept = FALSE)
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{coding}{Type op coding that need to be used. See ?contrasts for more information.}

\item{intercept}{Logical argument indicating whether an intercept should be included. The Default is False.}
}
\value{
A list containing 2 matrices, one contains all possible profiles with discrete levels, the other contains the coded version.
}
\description{
Function to generate all possible combinations of attribute levels (i.e. all possible profiles).
}
