% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{Profiles}
\alias{Profiles}
\title{Profiles generation.}
\usage{
Profiles(lvls, coding, intercept = FALSE)
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{coding}{Type op coding that need to be used. See \code{\link[stats]{model.matrix}}}

\item{intercept}{Logical argument indicating whether an intercept should be included. The default is False.}
}
\value{
A matrix which contains all possible profiles.
}
\description{
Function to generate all possible combinations of attribute levels (i.e. all possible profiles).
}
\examples{
# numeric vector specifing the number of levels for each attribute.
attribute.levels <- c(3,4,2) #3 attributes with respectively 3, 4 and 2 levels. 
coding.type <- "contr.sum" #effects coding will be applied.
#generate all coded profiles.
Profiles(lvls = attribute.levels, coding = coding.type)

# without coding
Profiles(lvls = c(4,4,3), coding = "none") 
}
