% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{Rdes}
\alias{Rdes}
\title{Random design generation}
\usage{
Rdes(lvls, n.sets, n.alts, coding, intercept = FALSE)
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{n.sets}{Numeric value indicating the number of choide sets.}

\item{n.alts}{Numeric value indicating the number of alternatives per choice set.}

\item{coding}{Type op coding that need to be used. See \code{\link[stats]{model.matrix}}}

\item{intercept}{Logical argument indicating whether an intercept should be included. The default is False.}
}
\value{
A design matrix
}
\description{
Function to generate a random design matrix.
}
