% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_algorithms.R
\name{SeqKL}
\alias{SeqKL}
\title{Sequential Kullback-Leibler based algorithm for the MNL model.}
\usage{
SeqKL(des = NULL, cand.set, n.alts, alt.cte = NULL, par.draws,
  weights = NULL, allow.rep = FALSE)
}
\arguments{
\item{des}{A design matrix in which each row is a profile. If alternative
specific constants are present, those should be included as the first
column(s) of the design. Can be generated with \code{\link{Modfed}}}

\item{cand.set}{A numeric matrix in which each row is a possible profile. The
\code{\link{Profiles}} function can be used to generate this.}

\item{n.alts}{Numeric value indicating the number of alternatives per choice
set.}

\item{alt.cte}{A binary vector indicating for each alternative if an
alternative specific constant is desired.}

\item{par.draws}{A matrix or a list, dependend on \code{alt.cte}.}

\item{weights}{A vector containing the weights of the draws. Default is
\code{NULL}, See also \code{\link{ImpsampMNL}}.}

\item{allow.rep}{Logical value indicating whether repeated choice sets are
allowed in the design.}
}
\value{
\item{set}{Numeric matrix containing the choice set that maximizes the expected KL divergence.}
\item{kl}{Numeric value which is the Kullback leibler divergence.}
}
\description{
Selects the choice set that maximizes the Kullback-Leibler divergence between
prior parameter values and the expected posterior, assuming an MNL model.
}
\details{
The algorithm selects the choice set that maximizes the Kullback-Leibler
divergence between prior and expected posterior. Otherwisely framed the
algorithm selects the choice set that maximizes the expected information
gain.
}
\examples{
# KL efficient choice set, given parameter draws. 
# Candidate profiles 
cs <- Profiles(lvls = c(3, 3), coding = c("E", "E"))
m <- c(0.3, 0.2, -0.3, -0.2) # Prior mean (4 parameters).
pc <- diag(length(m)) # Prior variance
set.seed(123)
ps <- MASS::mvrnorm(n = 10, mu = m, Sigma = pc) # 10 draws.
ac <- c(0, 0) # No alternative specific constants. 
# Efficient choice set to add. 
SeqKL(cand.set = cs, n.alts = 2, alt.cte = ac, par.draws = ps, weights = NULL)

# KL efficient choice set, given parameter draws. 
# Candidate profiles 
cs <- Profiles(lvls = c(3, 3), coding = c("C", "E"), c.lvls = list(c(5,3,1)))
m <- c(0.7, 0.3, -0.3, -0.2) # Prior mean (4 parameters).
pc <- diag(length(m)) # Prior variance
set.seed(123)
ps <- MASS::mvrnorm(n = 10, mu = m, Sigma = pc) # 10 draws.
ac <- c(1, 0) # Alternative specific constant. 
# Efficient choice set to add. 
SeqKL(cand.set = cs, n.alts = 2, alt.cte = ac, par.draws = ps, weights = NULL)
}
\references{
\insertRef{crabbe}{idefix}
}
