% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_support.R
\name{SurveyApp}
\alias{SurveyApp}
\title{Shiny application to generate a discrete choice survey.}
\usage{
SurveyApp(des = NULL, n.total, alts, atts, lvl.names, coding, buttons.text,
  intro.text, end.text, c.lvls = NULL, crit = NULL, alt.cte = NULL,
  prior.mean = NULL, prior.covar = NULL, cand.set = NULL, m = NULL)
}
\arguments{
\item{des}{A design matrix in which each row is a profile. Can be generated
with \code{\link{Modfed}}}

\item{n.total}{A numeric value indicating the total number of choice sets.}

\item{alts}{A character vector containing the names of the alternatives.}

\item{atts}{A character vector containing the names of the attributes.}

\item{lvl.names}{A list containing character vectors with the values of each
level of each attribute.}

\item{coding}{A character vector denoting the type of coding used for each
attribute. See also \code{\link{Profiles}}.}

\item{buttons.text}{A string containing the text presented together with the
option buttons.}

\item{intro.text}{A string containing the text presented before the choice
survey.}

\item{end.text}{A string containing the text presented after the choice
survey.}

\item{c.lvls}{A list containing numeric vectors with the attributelevels for
each continuous attribute. The default is \code{NULL}.}

\item{alt.cte}{A binary vector indicating for each alternative if an
alternative specific constant is present. The default is \code{NULL}.}

\item{prior.mean}{Numeric vector indicating the mean of the multivariate
normal distribution (prior).}

\item{prior.covar}{Covariance matrix of the prior distribution.}

\item{cand.set}{A numeric matrix in which each row is a possible profile. The
\code{\link{Profiles}} function can be used to generate this.}

\item{m}{Numeric value. Number of samples = \code{base^m}.}
}
\value{
A list object named "survey" can be found in the global environment.
\item{bin.responses}{A binary response vector indicating the observed
responses during the survey.} \item{responses}{A character string
containing the responses observed during the survey.} \item{design}{The
coded design matrix containing all presented choice sets during the
survey.}\item{survey}{All the choice sets (decoded), that were presented on
screen.}
}
\description{
This function starts a shiny application which puts choice sets on screen and
saves the responses. The complete choice design can be provided in advance,
or can be generated sequentially adaptively, or can be a combination of both.
}
\details{
A pregenerated design can be specified in \code{des}. This should be a matrix
in which each row is a profile. This can be generated with \code{Modfed}, but
is not necesarry.

If \code{n.total} equals the number of \code{nrow(des)} /
\code{length(alts)}, the specified design will be be put
on screen, one set after another, and the responses will be saved. If no design
is specified in \code{des}, \code{n.total} choice sets will be generated
adaptively sequentially using the KL criterion (see \code{\link{SeqKL}} for
more information). If \code{n.total} > (\code{nrow(des)} /
\code{length(alts)}), first the specified design will be shown and afterwards
the remaining sets will be generated adaptively sequentially.

Whenever adaptively sequentially sets will be generated, \code{prior.mean},
\code{prior.covar} and \code{cand.set} should be specified.

The names specified in \code{alts} will be used to label the choice
alternatives. The names specified in \code{atts} will be used to name the
attributes in the choice sets. The values of \code{lvl.names} will be used to
create the values in the choice sets. See \code{\link{Decode}} for more
details.

The text specified in \code{buttons.text} will be displayed above the buttons
to indicate the preferred choice (for example: "indicate your preferred
choice"). The text specified in \code{intro.text} will be displayed before
the choice sets. This will generally be a description of the survey and some
instructions. The text specified in \code{end.text} will be displayed after
the survey. This will generally be a thanking note and some further
instructions.
}
\references{
\insertRef{crabbe}{mnldes}
}
